+++ CHANGELOG.md
@@ -13,6 +13,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix playback of content with undefined colorspace
 - Fix NVIDIA color normalization with VP9 sources
 
+### Changed
+- Merge generated `Other Video` folder tags with tags from sidecar NFO
+
 ## [0.7.3-beta] - 2023-01-25
 ### Added
 - Attempt to release memory periodically